feat: update sheet table
