+++ frontend/src/views/sql-editor/Sheet/SheetTable/index.ts
@@ -1,0 +1,3 @@
+import SheetTable from "./SheetTable.vue";
+
+export default SheetTable;
