+++ frontend/src/views/sql-editor/Sheet/types.ts
@@ -1,0 +1,2 @@
+export const SheetViewModeList = ["my", "shared", "starred"] as const;
+export type SheetViewMode = typeof SheetViewModeList[number];
